♻️ Refactor: DetailModal 컴포넌트를 detailModal 폴더로 이동 (#17)
DetailModal 컴포넌트를 detailModal 폴더로 분리하여 구조 정리 및 import 문 순서 정리
